+++ src/specfem3D/noise_tomography.f90
@@ -424,7 +424,7 @@ end module user_noise_distribution
     if (NSPEC_TOP > 0) then
 
       ! check integer size limit: size of b_reclen_field must fit onto an 4-byte integer
-      if (NSPEC_TOP > 2147483646 / (CUSTOM_REAL * NGLLSQUARE * NDIM)) then
+      if (NSPEC_TOP > int(2147483646.0 / (CUSTOM_REAL * NGLLSQUARE * NDIM))) then
         print *,'reclen of noise surface_movie needed exceeds integer 4-byte limit: ',reclen
         print *,'  ',CUSTOM_REAL, NDIM, NGLLSQUARE, NSPEC_TOP
         print *,'bit size Fortran: ',bit_size(NSPEC_TOP)
